feat: First Draft of `FormObserver`
Also added some `Development Notes` for learning
(and remembering) important pieces of information.

Overall, the code here seems mostly functional.
We're committing it now so that we can perform more
iterations and look at the history. We have no tests
yet, but hopefully we'll have some in the near future.