refactor: Make `FormObserver` _Interface_ Non-generic
Although the constructor of `FormObserver` needs to be
generic, the individual methods that the class will use
will always look the same. Therefore, there's no reason
to require the `interface` itself to be generic. We
have updated our code accordingly. (This also lets us
remove the ESLint disable comment as a bonus.)