chore: Enforce Implementation of `FormObserver` Interface
The most important thing we need to do is have good
constructor overloads for our end-developers. BUT,
if we can ALSO add some additional type safety to
our own code while we develop, that's an added
bonus. This doesn't seem to mess anything up, so
we'll keep this for now.