fix: Correctly Infer the `listeners` Tuple Type for `FormObserver`
Previously, for the 3rd constructor of `FormObserver`,
TypeScript was attempting to use `listeners` to _infer_
the type of `T`. If explicit argument types weren't
provided for the functions in `listeners`, then TS
would consequently generalize `T` to `ReadonlyArray<EventType>`
for `listeners`. (Recall that `ReadonlyArray<EventType>`
is the constraint of `T`.)

Now, the `listeners` type is based on `U`, not `T`
(even though `U` is `infer`red from `T`). Therefore,
TS no longer tries to infer `T` from the `listeners`
argument of the 3rd `FormObsever` constructor overload.
Instead, TS -- having already determined `T` from `types`
(which is now its only way to determine `T`) -- uses
`T` to determine a separate type `U`, which in turn
is used to determine the type of `listeners`.