feat: Enable `FormObserver` to Indicate If a `Form` Is Already Being Observed
`FormObserver.observe()` now returns a `boolean` that indicates
whether or not the `form` it received was already being observed
when the method was called. The `FormObserver.unobserve()` method
does something similar. This will make life vastly easier for
anyone trying to extend the `FormObserver` class and reuse some
of the information/logic in this class.