feat: Add Full Test + Acceptance Coverage for `FormStorageObserver`
- Test coverage for `FormStorageObserver` is 100%.
- From what we can tell, "Acceptance Coverage" for
  `FormStorageObserver` is 100%.
- From what we can tell, TypeScript types "test
  coverage" for `FormStorageObserver` is 100%.
  (This DOES NOT interfere/interact with the true
  Jest test coverage.)

Additional Changes:
- Added some notes about things we need to inform the end
  developer about when it comes to using
  `FormStorageObserver`.  We should address this information
  next before moving on to the next part of our project.

NOTE: We also changed the way that type narrowing occurs
during data loading for multiselect fields. This was
done to resolve a _false negative_ that we saw in our
code coverage. After making this change, we realized
that this check is actually helpful for guaranteeing
that the user doesn't accidentally try to load
non-array data into their multiselect element (if
they somehow corrupted or failed to update/clean the
data they had in `localStorage`). This is important
because it prevents our code from breaking if they
load an invalid value. :) Nonetheless, we don't
CURRENTLY think that this guard is something that
should be added to our Tests or to our Acceptance Criteria.