fix: Correct `FormStorageObserver.load` Error Messaging
We recently discovered that if one field has an `id`
(or `name`) that matches the `name` (or `id`) of another
field, the current JavaScript standard will mistakenly
group these fields together under a `RadioNodeList` EVEN
IF the two fields are NOT radio buttons. This means that
the result of `form.elements.namedItem(ID)` and
`form.elements.namedItem(NAME)` for the aforementioned
elements will result in a `RadioNodeList` being received
instead of the proper input fields.

There's not really a (clean, performant) solution to
this problem. So we're removing this scenario from
the list of errors/hints that we provide, we're adding
a note to create documentation that helps people
avoid this pitfall, and we're improving our errors/hints
for the cases where we _can_ better steer the end
developer towards code that functions properly.