WIP: Correct Docs for `FormValidityObserver.validateFields`
We've concluded that it doesn't really make sense to
validate all of the fields if the developer (technically)
specifies that they want _no_ fields validated by passing
an empty array. If someone wants to validate all the fields,
they can omit the argument. If someone wants logic that
will allow them to validate 0 fields in some cases, we will
not prevent them from doing so.

Advances the Following WIP(s):
- 1cc4808becdbb306596eb316b46dee0577ae3766