WIP: Enable `validateFields` to Work with Unregistered Fields
After we discovered that fields _don't_ need to be
registered in order to be validated, we updated
our code accordingly. However, `validateFields` was
not updated to handle this reality. We have now
corrected `validateFields` to work even if no
fields have been registered. (Note, we are
are talking about the "overload" where no arguments
are passed to `validateFields`. The implementation
for when specific field names are passed to
`validateFields` has remained unchanged.)

From a performance standpoint, we're hoping that this
implementation should be efficient for what we're
trying to accomplish. However, for those who want
to micro-optimize, we'll consider adding additional
optimization features in the future. (We'll only
do this if we find that there's a meaningful
performance benefit to developers that comes from
this. We don't want to waste time. And what we have
seems pretty good and practical.)

Advances the Following WIP(s):
- 1cc4808becdbb306596eb316b46dee0577ae3766