WIP: Use a `Map` for `FormValidityObserver`'s Stored Error Settings
Our understanding is that a `Map` is more appropriate
and more performant than a raw object when working with
dynamic key-value pairs. (The pairs are dynamic because
the end-developer can register any number of form fields
and the key-value pairs must all be cleared when the
`FormValidityObserver` unobserves a `form`.

Advances the Following WIP(s):
- 1cc4808becdbb306596eb316b46dee0577ae3766