chore: Add Documentation about Web Components Support
If possible, we'll want to figure out a way to run
automated tests for our Web Components support. JSDOM
doesn't currently support `ElementInternals`, so we
probably can't do this with Jest.

It might also be sufficient to just test the Web Components
support manually with the demos that we expose in the
future (given the limitations of automated testing tools).