WIP: Improve `register` and Achieve Full Test + Acceptance Coverage
We have simplifed the `FormValidityObserver.register` method
to make it more DX-friendly and more performant. You can
read up on the reasoning behind these decisions in the
`DESIGN_DECISIONS.md` file.

As with before, we have accomplished full test and
acceptance coverage for this method. (In this case,
the test coverage and acceptance coverage are pretty
much one and the same.) Improvements may come in the
future, but this method is so simple that there likely
won't be that much to do with it.

We also added some comments to `FormValidityObserver.ts`
that could be helpful for improving the class in the future.

Advances the Following WIP(s):
- 1cc4808becdbb306596eb316b46dee0577ae3766