chore: Add Type-Only Tests for `FormValidityObserver`
We forgot this important part of testing. We need to
make sure that the API which our users experience
is consistent between updates to the package (assuming
we're trying to avoid breaking changes). Now we've
added the missing tests.

When the `FormValidityObserver` receives new options,
these tests will need to be udpated.