chore: Delete Unused `FormValidityObserverV2.ts` File
Originally, we kept this file around so that we could keep
track of our original approach to the `FormValidityObserver`.
Since we've completed the first draft of the
`FormValidityObserver`, this file is no longer necessary.
It was only around in case we wanted to "go back" or
use it as a reference.