refactor: Remove React Testing Library from All Core Library Tests
Yep. We're this devoted to "pure JS" for the core. In the
JS-framework tests, we'll use the framework-related testing
tools. This was a bit painful for the `FormStorageObserver`
though.

In situations like this, having our over-paranoid assertions
before/within our tests is actually REALLY HELPFUL for
guaranteeing that we didn't break anything during a test
restructure. Hopefully we won't have another overhaul
like this in the future.