chore: First Draft of `FormValidityObserver` Docs (without Full Code)
The docs aren't perfect. But again, we're trying to get
_something_ that's readable done so that we can release
the package. Then we can put more effort into the docs
later (especially if the package ends up being useful).
This might be a risky move since docs are important
for adoption; but given our current time constraints,
we don't have a lot of options here.

- We still need to add code examples for the main docs.
- We have some TODOs to address regarding additional
  potential documentation content. However... we're
  debating if we need to write those things before release.