chore: Resolve Pending Docs Comments in the `FormValidityObserver`
Here's how each Documentation Comment was handled:

1) Warning Developers about Error Message Ordering

In practice, browsers are already inconsistent with how
they order error messages, so there isn't really a need
to warn developers about what they've already been
working with.

Instead of adding something to the primary documentation,
we added some small `Development Notes` regarding why
we chose the ordering we did for error messages, and when
this ordering is overriden by the browser.

2) Explaining Why We Simultaneously Support Accessible/Native Errors

We concluded that this isn't really something worth
explaining. It doesn't really seem like documentation that
would help developers understand or use our library better.
So we've foregone the explanation altogether and deleted
this comment. There are enough development/design notes
lying around to explain the significance of being able to
leverage `field.validity.valid` when needed by taking
advantage of `setCustomValidity`.

3) Handling Non-Bubbling Events

We also forewent adding this to the `FormValidityObserver`
docs. In practice, the `invalid` event would not be
dispatched until _after_ a field was already validated,
so no one will be calling the `FormValidityObserver`
constructor with the `invalid` event type to begin with.
The other more important event types have bubbling
alternatives, such as `focusin` and `focusout`.

Moreover, there are other parts of the documentation
that show how event capturing can be leveraged. So it
just doesn't seem that necessary to dive into the
details of event capturing here.

4) Explaining Our Validation Methods as Enhancements of the Browser's

Our documentation is already doing this, so the comment can
be safely deleted.

5) Explaining Why We Don't Use State for Error Messages

We haven't decided whether or not we should have a
"Philosophy" section of our documentation yet. So, for
the time being, we've put an explanation for this in
the `Development Notes`. Perhaps in the future we can
add a philosophy page... if people would even be interested
in it.