feat: Add Svelte Integration + Related Docs (`FormValidityObserver`)
NOTE: The `createFormValidityObserver` utility has a dangling
TS error related to how we import everything. We'll take
care of this later. (This is a CJS vs. ESM issue, but around
TS types, oddly enough.) The functionality at least seems
solid.

Docs are written. A new workspace has been added. And
potential items to address in the future have also been
pointed out.

We haven't listed `svelte` as a `peerDependency` yet because
`svelte` technically isn't needed. All that we have to do
is define a single function and we're good to go. We only
"need" `svelte` for the explicit TypeScript types; but no
runtime logic is required from `svelte`. Consequently,
we'll probably leave this as is. Svelte really is an
amazing framework.