chore: Add Tests for `createFormValidityObserver` (for Svelte)
NOTE: In order to run these tests, you'll have to replace
Svelte's type definitions for `HTMLInputAttributes` with
a custom type definition local to the file. For some
weird reason, TypeScript considers the import of the `.d.ts`
file to be an ESM import... so we can't import it and run
our Jest tests. Bothersome, to say the least. There are
some other problems that surround using Svelte + Jest...
see the testing file for more details.

In addition to adding tests, we made some minor updates to
the `createFormValidityObserver` (and its documentation).

We corrected the Jest configuration. And we _started_
to get Svelte's configuration up and running for Svelte.
However, as mentioned earlier, there are complications
here; so we'll have to address this later -- after release.