refactor: Migrate to ESM and Vitest
CommonJS --> ESM
Jest --> Vitest

We're basically doing this because of the headaches
we encountered with Svelte being an ESM. This change
brings new headaches, but the DX of `vitest` does
seem to be better than that of `jest`. We'll see
how this goes.

ESM probably would've come around to this project at
some future point anyway? So this is just something
that we're taking care of early -- though not with
much excitement.