refactor: Use JSDocs Directly for `FormObserver` and `FormStorageObserver`
This allows us to have our documentation and our types
in the same place! Hopefully it'll make developer experience
better in the long run. We'll make this update for the
`FormValidityObserver` as well.

HUGE thanks to @DanielRosenwasser! (See his current
solution for JSDocs in microsoft/TypeScript#55916.)

This is GAME CHANGING.