refactor: Only Use JSDocs for `FormValidityObserver` (BREAKS TYPES)
As we mentioned earlier, this makes life much easier.
Unfortunately, this change breaks the types that were
already illegal. (This is technically a good thing.)
But we really wanted to push this change through.
We'll address the broken types later.

The main types "breakage" that occurred is that using
`FormField` for the `validate` function and for
`ErrorMessage<any>` functions doesn't work at all
anymore. But that type was technically already broken.
Now we're just being forced to resolve the issue,
which we already noted in our "Things to Do
before Releasing" notes.

So far so good! Tests are still passing, of course.
And even better, our types and our documentation
and our implementation are all cleanly in one place
now.