fix: Update Svelte Form Validity Observer Types to Work with Core
ALSO: Updated the `autoObserve` method to take a
`novalidate` option that indicates that the "novalidate"
attribute should be applied to the `form` element during
the mounting phase. (Defaults to `true`.)

NOTE: There are some `package.json` and `tsconfig.json`
updates for the `core` package that we have been excluded
from this commit. If anything related to the `core`
package is broken in this commit, that's why. Look
for the next commit that updates/corrects these files
if needed.

Note: There are some weird things about `vitest` types
that might be worth addressing later. But that isn't
urgent since it's only related to the test files.