fix: Update React Form Validity Observer Types to Work with Core
ALSO: Updated the `autoObserve` method to take a
`novalidate` option that indicates that the "novalidate"
attribute should be applied to the `form` element during
the mounting phase. (Defaults to `true`.)

Note: There are some weird things about `vitest` types
that might be worth addressing later. But that isn't
urgent since it's only related to the test files.

Also made a slight improvement to svelte integration docs.