chore: Establish Exports and TS Configs for Packages
It seems that this is an important part of getting
ready for the build process. Note that we'll still
need to figure out how we'll support CJS. But we
might do that _after_ we've figured out how to
release an ESM version.

Should resolve the potential concern mentioned in
d00e67fdbe591b61214ca80f91f2988dcf13d988.