chore: Create Build Step for Generating Types + CJS Files
We might iterate on this in the future, but this
seems to do the trick of properly generating
the `.d.ts` and CJS files. Next we need to
test publishing.

It's a little annoying that we still need _some_
kind of build step, though. :\ But I guess it's
simpler than working with pure `.ts` files.