chore: Update Features and Benefits of Form Observer
We also concluded that we won't mention that we
support TS types in our features and benefits.
I think in today's day and age, it's assumed that
your library simply has good types if it's worth
using. Conform and React Hook Form don't mention
anything about TS types, it seems. So... it's
probably assumed that types will just work for
our library.