feat: Add Convenience APIs for Solid
Includes documentation and tests. Note that
the tests for Solid are a little busted right
now because there seem to be bugs either in
`vitest` or in `@solidjs/testing-library`.
It's unclear right now. But this isn't something
that we're willing to let block our project; so
we'll just keep pushing on and address this in
the future. Honestly, the testing problem is
a little disappointing. :pensive: