chore: Indicate Bundle Size of Project (gzipped)
To gather this information, we leveraged the
`vite-bundle-visualizer` NPM Package with a simple
Vite App.

To quantify the bundle size of JS Framework Integrations,
we used a Solid + Vite App because its bundle size will
generally be larger than that of the other integrations
(due to the custom error renderer function for Solid).