refactor: Drop Support for `setHTML` (FormValidityObserver)
Surprisingly, it seems that as of today, many browsers
have _dropped_ support for the `setHTML` method. I guess
something went wrong? So... for now, _we_ are also dropping
support for it. (The method also acts buggy in some cases.)
Once it is better supported (and properly, permanently so),
we can add the use of this method back in.