feat: Clarify That Packages Have No Side-Effects
For some reason, we have to point this out in
our `package.json` file even though our libraries
are truly side-effect free. It seems that
React Hook Form had to do this as well, so maybe
this isn't so unusual?

After playing with the Rollup REPL and using an
example similar to our `FormValidityObserver` from
the `core` and our `useFormValidityObserver` from
the `React` integration, we concluded that
everything in each of our libraries is properly
tree-shakable. (The only possible candidates for
side effects are the `FormValidityObserver`
and the `ReactFormValidityObserver` due to
locally-declared variables. But as we saw with
Rollup, the way that we're doing things results
in 0 side-effects.)