fix(core): Enable the `FormObserver` to Work with ShadowDOMs
Our previous approach to event delegation was a bit naive.
Now we can properly support `form`s in the ShadowDOM
whether they're `open` or `closed`. Of course, the
restrictions that we've mentioned elsewhere about
not mixing fields between the Light DOM and the
Shadow DOM (or between various ShadowDOMs) still
holds. That is a Web Standard enforcement that
cannot be undone.