fix(core): Enable the `FormValidityObserver` to Work with ShadowDOMs
We forgot that the `FormValidityObserver` had references to
`document` instead of the "True Root Node". So now we've
fixed the `FormValidityObserver` too.

Note that the `FormStorageObserver` has never had any
references to `document`, so it shouldn't experience
any issues with the Shadow DOM since the core
`FormObserver` class was already fixed.