chore: Add Docs for Handling Form Data and Conditional Fields
Since there are other tools that allow developers to
convert form data to JSON on the frontend, it's important
to equip developers to do the same thing on the
frontend or backend as they please. It's also important
to point out why using JSON may not be the most
advantageous approach.