refactor(core): Make the `FormObserver` Track Relevant Root Nodes
The `FormObserver` has been updated to keep track of the
Root Node that _originally_ owned an `HTMLFormElement`
when it was observed. This enables us to properly
remove event listeners even if the `form` was removed
from the DOM (or moved to a new Root) before `unobserve`
was called. In the end, this basically helps protect
developers from accidentally creating memory leaks.

The `FormObserver` also refuses to re-observe a `form`
that is moved to a new Root until it is first unobserved.
Again, this helps developers avoid unintended memory
leaks.

This change was made because `svelte@4` and Lit do
not give developers an easy way to access the original
Root Node of a `form` during its teardown process.
Our new approach technically takes some flexibility
away from developers, but it also makes memory safety
more easy/automatic.