refactor(core): Use `renderer` for Clearing Error Messages
Apparently, many JS Frameworks/Libraries NEED full
control of the DOM Node into which they render content.
So, for instance, if a framework renders "ERROR" into
an error container, it ALSO needs to be responsible for
clearing that error container. Otherwise, if some outside
JS logic interferes with the container's content (e.g.,
`element.textContent = ""`), then the framework just
just doesn't know what to do during rendering anymore
(at least for Lit and Preact).

By calling the `renderer` with `null` when a field
becomes valid (or when `clearFieldError` is called),
we become able to more fully support custom renderers
AND regular error message strings.

Note: In the future, we may explore rendering a
constant `Symbol("empty")` instead of relying on
`null` if that makes sense. But up until now, most
reputable frameworks/libraries have accepted `null`
as meaning "Erase the element's content!" This is
worth leveraging because it means that developers
will have to put in less work when defining custom
`renderer`s that simply leverage the `render` function
of their framework/library.