chore: Forego Creation of Qwik Integration (for Now)
After updating our Core Package to support the Shadow
DOM in a more robust and performant way, we found
that our tool was no longer able to integrate into
Qwik in an _intuitive_ way. This is because of the
actions that Qwik takes to ensure resumability.
Basically, our tool won't work outside of a
Qwik Task anymore. And the solutions that adhere
to these restrictions can result in a less
intuitive developer experience. I guess this
outcome isn't too surprising, given that Qwik
itself is not very intuitive (e.g., devs not being
able to access `event.currentTarget` or call
`event.preventDefault()` from within event
handlers as usual).

We don't have much interest in creating an
integration that doesn't have an intuitive
developer experience at the moment. Doing so
would require more time than we currently are
willing to sacrifice. And this time would be
sacrificed for a tool that hasn't quite risen
in popularity like the other frameworks (and
that may remain less popular because of other
alternatives that are more intuitive and that
give apps the same speed simply by minimizing
the amount of JavaScript in an app altogether).

We've saved the progress that we had originally
made on the `unstable/previous-qwik-efforts`
branch. If we decide to pick up where we
left off and make a working integration for Qwik,
that branch will be the starting point.