chore: Don't Type Check `.d.ts` Files
The TypeScript type checker is now throwing
errors when we check types with the
`skipLibCheck: false` configuration. It seems
that one of the packages that we installed along
the way started causing this problem... Vitest
definitely seems to be a culprit. We can't
ignore this anymore since we're thinking about
adding linting to CI.

Generally speaking, this should be fine. But
we DO want to make sure that our LOCAL `.d.ts`
files are accurate. So... if necessary, we may
occasionally need to make this flag `false`
to make sure things are behaving as expected --
unless we find that TS will still check our
LOCAL `.d.ts` files.

In an ideal world, we could define all of our
types in JS files. Hopefully that day comes
eventually. Then we wouldn't need to worry
about this problem either way. :smile: