refactor: Only Allow 1 Event Type to Be Specified for the `FormValidityObserver`
Although this is a breaking change, it will more easily
allow us to support the `revalidateOn` feature that
we mentioned on GitHub while also maintaining the
parent-child relationship between the `FormObserver` and
the `FormValidityObserver`. (We could technically choose
to break that relationship, but the reusable `observe`
and `unobserve` methods seem to be proving helpful right
now.)

This change also seems more logical because -- practically
speaking -- developers are likely going to choose one
event anyway. They'll have to choose an event that actually
relates to user interaction. And the most common of those
events are `input`, `focusout`, and `change`. However,
the final `input` triggers before `focusout` happens, so
it doesn't make sense to specify both simultaneously.
Similarly, the `change` event will only be relevant for
a subset of the times that the `focusout` event is
triggered, so supplying those two simultaneously also
doesn't make sense. So developers will likely only
supply 1 event at a time.