chore: Deprecate `useFormValidityObserver` React Hook
Originally, this hook seemed like a good idea. But actually, it's
turned out to be something less performant than expected, and
something that's more prone to footgunning than expected. See the
`Design Decisions` document for additional details.

Notice that `react` is still a peer dependency of `@form-observer/react`.
In the future, we might use the React package to spy on the
React version so that we can figure out how to appropriately
generate the `autoObserve` function. If we decide against that,
we can remove the peer dependency. Again, see the `Design Decisions`
document for additional details.