chore: Update TODOs (for Lit and Package Version)
Note: We actually didn't change the default `renderer` for the Lit
integration. That would be a breaking change, and we don't know if
developers would even want that to be the default behavior yet. So
instead, we updated our StackBlitz example to show how to use a
different kind of `renderer`. We also added a stateful example.