chore: Apply Minor Package Updates and lock `vite-plugin-solid`
For whatever reason, different versions of `vite-plugin-solid`
seem to cause different kinds of troubles. So at this point, we just
need the lock the library where it's at. Hoepfully we can find out
what the underlying problem is later, but it seems to be a problem
with the Vite Plugin.